Reformat logging with timestamps.
Print timestamps, module, and log level in format that stays (mostly)
in a fixed format for alignment. To get timestamps without having to
constantly check the time, we track the last time value from calling
Timer::now() and a process-wide epoch value so we can report in
seconds rather than an absolute date-time.