Disconnect presences that finish connecting after HostedObject destruction.
The SessionManager may finish connecting an presence after the
HostedObject that requested the connection has been stopped or
destroyed. Previously we just ignored this, but that can cause us to
hold onto the object as it still has presences. Instead, when we
detect this, force a disconnection request by the ObjectHost so the
presence is cleaned up.

This fixes an intermittent issue with the Emerson system storage
test. Sometimes the test would finish before the automatic connection
process finished, which would result in killEntity getting called
while a connection was in process. The test would then hang and
timeout. Now it exits cleanly.