Make sure input events are properly paired.
Since a single input event stream is used across both WebViews and
Invokables (scripts), we can get into cases where the start of a
logical sequence (e.g. mouse down, mouse drag, mouse up) is sent to
one component and the rest is sent to another. This can be due to
unrelated events, e.g. a mouse click on a web view while a key is
pressed.

To fix this, add a simple class that tracks these in-progress logical
events and the current target. When the target changes, generate fake
events to complete the logical sequence and then continue sending
input to the new target.

Fixes #341.