Use Boost 1.46 for Debian build.
Previously, 65b9c1 switched to using Boost 1.48 because of an issue
with Boost.Math in Boost 1.46. However, Boost 1.48 has a more serious
issue with option parsing which breaks (Boost issue 6264). Instead fix
the comparatively more minor issue with Boost.Math in f0920e and
switch to Boost 1.46 in the Debian control file.