Use simplified convex hulls for dynamic per-triangle physics.
Change computation of collision shapes to intelligently select between
a BVH of the full set of mesh triangles for static objects and a
simplified convex hull of the input mesh for dynamic objects. This
allows "per-triangle" physics for dynamic objects, where previously
setting this would result in a BVH shape which was just ignored. Fixes