Use a less garish theme.
Add the "Vader" jquery UI theme and switch to it. It's a lot more
toned down than the blue theme we already had. Leaving the old one in
in case we want to provide theme switching. This also simplifies the
JQuery Menu css and uses the JQuery classes instead of specifying
colors directly so that the menu will get properly updated when we
switch between JQuery themes.