Refactor interaction with top-level Pinto into LibproxProximityBase.
Make LibproxProximityBase responsible for interacting with top-level
Pinto, only requiring a few calls from implementations and then
informing them as servers become relevant or disconnect. This
functionality should be the same for all implementations so it's
better provided as a simple service in the base class.

With that new base-class functionality, make LibproxManualProximity
"handle" notifications by logging and ignoring them since it isn't
setup for server-to-server queries yet.