Tag loc updates with prox indexes.
Add a prox index field to the Loc update protocol. When replicating
indexes (trees), the Proximity implementation usees a new version of
the subscribe/unsubscribe interface to LocationUpdatePolicy which
includes the index that caused the subscription. The old-style
interface still works by using an assumed single, implicit index which
isn't actually tracked or reported. The update policy then reports the
list of indexes for the observed object and observer so the observer
can apply the update to the correct indices.

The manual_query plugin on the OH uses this information to avoid
applying the update to all indices but still get the orphan updates to
apply properly since it knows the update needs to be applied to an
index even if it wasn't aware that index exists yet.