Refactor master-pinto code to support solid angle and manual queries.
Refactor the master-pinto code, which used to be a single simple
class, into three classes, much like the proximity code. The base
class handles basic connectivity to the server and the subclasses only
need to deal with sending query updates and receiving results. This
also adds the class for handling manual queries, but currently it's
just a clone of the aggregate solid angle query version since we need
pinto server support for manual queries as well before we can
implement manual querying.