Refactor PintoManager to prepare for addition of manual PintoManager.
Refactor PintoManager into a base class and a specialization that
deals witha specialization. The base class handles the basic
connections and the implementation performs the actual querying and
handles query-format-specific messaging, i.e. anything besides the
basic tracking of ServerID, server region, and largest object.

In doing this, the format of pinto query updates was changed from the
solid angle query PBJ message to a simple byte array so the interface
in PintoManager could be set to its final form. This required updating
the existing clients of PintoManager, the MasterPintoServerQuerier
classes, to serialize a QueryUpdate message into this field instead of
embedding it directly as used to be possible.