Factor ManualReplicatedClient out of ServerQueryHandler.
Refactor ServerQueryHandler, moving the core functionality of managing
the remote replicated tree and pushing the data into a
LocationServiceCache into a separate, reusable class,
ManualReplicatedClient, in libpintoloc.

ManualReplicatedClient deals with handling prox and loc messages,
resolves orphan loc updates, and then generates output in the form of
LocationServiceCache entries. It also accepts input specifying which
nodes are being observed so it can decide when to refine the tree.

The user of the class, ServerQueryHandler in this case, is responsible
for injecting loc and prox events, information about which nodes are
observed. It also gets events indicating when new trees are replicated
(since there may be multiple trees for a single server,
e.g. static/dynamic or rebuilding) and for sending (opaque) commands
to the server to control the cut.