Use ManualReplicatedClient to make LibproxManualProximity interact with ManualPintoManager.
Add ManualReplicatedClient based on space server IDs to
LibproxManualProximity, currently using only NullServerID to handle
events from the PintoServerQuerier. This triggers the creation of a
ManualReplicatedClient and a few helper methods complete the basic
messaging, forwarding query update requests to the
ServerQuerier. Currently only the initial exchange is happening
because a query handler isn't yet being used, meaning usage on the
nodes is not reported.

Other changes were necessary to make this work. Duplicate
ProxSimulationTraits files were removed, preferring a single copy in
libpintoloc. A NopTimeSynced implementation was added to be used when
no time sync is necessary or available. PintoManager now needs
aggregates, so PintoManagerLocCache was updated to support those,
PintoManagerBase implements the libprox AggregateListener interface
and forwards the events to PintoManagerLocCache, and
ManualPintoManager sets itself up as a listener for aggregate
events. ManualPintoManager also needed to report a bit more
information with results -- a timestamp that is required in the pbj
definition and IndexProperties which the client uses to know when to
create a new index.