Make all query processor classes, including the PintoManagers, provide a base set of commands.
Add BaseProxCommandable class to register a base set of commands for
query processor classes to handle and define pure virtual handlers for
them so subclasses must implement the commands. Use this to simplify
registration of these commands on the object host
(ObjectQueryProcessor), space (Proximity), and pinto
(PintoManagerBase). Add the implementation of these handlers for the
PintoManager classes.