Properly cleanup DownloadPlanner in OgreRenderer.
Make OgreRenderer properly stop and delete the DownloadPlanner. In the
DownloadPlanner classes, make sure Liveness::letDie is called and that
base class implementations of Service start and stop are called, all
the way down to PollingService, so that Timers are properly
cleared. This makes it so the DownloadPlanner shouldn't block clean
shutdown. For example, this makes meshview shutdown cleanly (other
binaries are still blocked by other things).