Fix timer cleanup in SQLiteStorage.
When stop was invoked, SQLiteStorage was just cancelling its timer,
then joining its IO thread. This could frequently end up blocking
because the other thread had no check for whether stop had been
called, so it could start reusing the timer, which would continue to
block the IO thread and in turn block the thread calling stop.

This changes the behavior to actually reset the shared_ptr so the
is actually removed and can't be reused. Where the timer is used we
need to copy the shared_ptr to ensure it stays valid, then call
wait().