Remove UUIDProxSimulationTraits.
This was leftover because the space server was originally written
around UUIDs rather than ObjectReferences. However, it'll cause
problems in reusing some of the libpintoloc code, which needs query
processors to use the same types. This removes
UUIDProxSimulationTraits, adjusting the libprox space plugins to work
with ObjectProxSimulationTraits and performs the necessary
translations between UUID and ObjectReference since the Proximity
interface and some others from the space server still use UUIDs.