Fix directing of query requests to the correct handler/query.
Query requests were not going to the correct query
handlers/queries. This happened because we were using the wrong index
IDs for identifying the queries. All the data was indexed by the
ProxIndexID from the replicated tree, but then the ProxIndexID for the
manual query handler that processes queries over the replicated data
was reported back to the client. The client would then use this ID in
the request, but we would interpret it as one of the remote IDs. This
actually "worked" frequently in testing in that it found a query/query
handler to work with because both local and remote indices would start
with 1, and it would just fail to refine/coarsen because no such
object existed in the incorrectly found tree.

This patch changes two indexes that we maintain: one now maps a local
ID to a ServerID + remote index ID pair instead of just mapping to
ServerIDs. The map from aggregators to indices also maintains all the
information it needs (server + remote index ID) instead of relying on
a secondary lookup.