Fix double notification of new index in ManualReplicatedClient.
The server can't tell perfectly from the results whether the index is
new because a lone addition of a node may also just be a new
root. Make ManualReplicatedClient handle this case by deferring
creation of the ReplicatedLocationServiceCache until it sees the index
properties in a message and check whether it's actually the first time
the index was seen before invoking the notifaction callback.