Fix thread-unsafe modification of meshes.
Move filtering of meshes (triangulate + center) to the point they are
loaded instead of right before they are used. The previous approach
wasn't thread-safe because the filters modified the data in place. The
new approach has the added benefit of only performing the filtering
once, so it is more efficient.