Add server-to-server queries for LibproxManualProximity.
Change query handlers over local objects in LibproxManualProximity to
be called "local handlers" rather than "OH handlers" since they can
also handle server-to-server queries, where only local objects should
be returned. Add support for server-to-server queries, which is mostly
a mix of code similar to that in LibproxProximity for basic message
handling setup, message queuing, polling, and sequence numbers, and
code similar to the existing OH query handling for local objects,
which was special cased. All queries share a single queryHasEvents
method which customizes and dispatches the events for the correct type
of client.

With these changes, the basic functionality for getting query results
from anywhere in the world to any querying object is complete.