Work around unreliable server-server messages.
There's only one location where server-to-server messages can be
unreliable, right after they're received and pushed into a shared
queue to cross threads. Unfortunately a number of server components
really rely on message reliability and ordering. While a real
reliability layer would be preferable, handle this for now by
whitelisting everything but ODP messages to be foreced into the
queue. As long as we trust server components to behave nicely then
this should be fine for now.