Fix IOStrand to ensure ordering of dispatch/post.
IOStrand is supposed to ensure that any dispatch or post calls
are executed with the same order in the strand, although wrapped
handlers do not have that guarantee. Because the implementation
had been using wrapped handlers, the ordering was not
guaranteed. In fact, it was frequently violated if invoked in
quick succession, but it just happens to matter rarely enough
that it wasn't frequently causing problems.

This switches to use the underlying ASIO method corresponding to
each call. When tracking queues, we need to change how we're
ensuring both the strand and service track the handler since we
must work through the underlying strand now. This also adds a
test to make sure the ordering constraint isn't broken again.