Track objects while they are in QueryEvents to ensure all events can be fully reported.
Update libprox to a new version which has an updated
LocationServiceCache interface which lets it make QueryEvent objects
force the LocationServiceCache to hold onto objects during the
lifetime of the QueryEvent, ensuring that recipients can get the data
to report results. This requires updating all the
LocationServiceCaches to support this interface and a few minor
changes where we also create our own QueryEvents. It should remove all
the cases where we had to ignore results due to them no longer being
in the LocationServiceCache.