Ensure coarsening caused by server doesn't remove buffer of refined nodes in manual replicated client.
Track observed nodes in ManualReplicatedClient and use them to
determine when a child of an observed node was removed. When one is,
issue an additional refine request. This covers the case where the
server needs to coarsen the cut (e.g. lifting for some operation) but
it doesn't end up affecting any cuts within our local copy of the
tree. Since that could leave us without a buffer and nothing to cause
a refinement request to go out we previously had to use any additional
(beyond the first) observer to trigger queriersAreObserving calls, but
with this fix we can remove that hack.