Use two mutexes, one for listeners and one for data, in CBRLocationServiceCache.
Using just one lock meant that all calls to listeners, even when they
were running in the right strand, had to be protected by the same lock
as all the data, mainly because they might register/unregister from a
differen thread. Because of this, we could end up with very long
delays as something like a connection event would spend a lot of time
doing work in a callback, blocking all access to the
LocationServiceCache. Since the callbacks don't actually need to
access the data directly, we can just use a different mutex for the
list of listeners.

The problem initially became obvious after the lock in
processObjectAdded was added in
f8fe2d88b14678140950d5889d59f7c85bc56968 because this caused other
operations to block waiting for objects to be inserted in a large tree
which might require a lot of changes to the tree.