Only register server queries if parameters actually allow results.
When receiving a server query update in LibproxProximity, only
register a query with the query handler based on the request if the
parameters allow results to be returned, i.e. if the query angle isn't
SolidAngle::Max. Also remove queries if they were active but now can't
return results.

This change is important because it keeps us from evaluating useless
queries and generating unnecessary traffic. This is actually important
because even though the query parameter suggests nothing should be
returned, objects within the query bounds will still be returned, and
for servers this can be a substantial number of objects.