Fix limit on outstanding loc update messages.
af526c0a49e9ab933e47d97938f3e1bc9747fd1 tried to fix the counting of
loc update messages being that were in the process of sending by using
the reference count on a shared_ptr. However, it used the wrong
shared_ptr, one which wasn't passed through callbacks, so it wasn't
using the correct count and could end up having any number of
concurrent outstanding messages. This just changes the function to use
the correct shared_ptr and requires a bit of rearrangement since the
method used to be inside the object we need to get the refcount of.