Fix marking of dirty aggregates.
There was an error in the way we marked dirty aggregates where we
both didn't mark an aggregate as dirty and stopped walking up the tree
if it had no children. Because in some cases we start from objects,
which by definition cannot have children, this failed to walk up the
tree and mark objects as dirty.

This changes the behavior to only mark non-leaf objects as dirty, but
do so regardless of whether they have children, and always walk up the
rest of the tree, potentially marking parents as dirty.