Add aggmgr.skip-upload flag.
The flag makes the upload step in the AggregateManager effectively a
nop, always indicating success. This is useful for testing without
requiring a CDN with OAuth credentials. It is disabled by default. For
large enough scenes it will fail because the eventually meshes will be
evicted from the cache and fail to download, but it should be fine for
testing against smaller scenes.