Fix missing creation of write callbacks in proximity streams.
There are two locations we can instantiate the ProxStreamInfos for
ObjectHosts and Objects. One of them was missing the creation of
callback which requires a weak_ptr to the created ProxStreamInfo, so
has to be created immediately after instantiation. This fixes that and
also notes that the two locations need to match.