Switch to using Boost's official UUID class.
The copy of the UUID class we had copied in from before it was an
official Boost library turns out to have not have generated
sufficiently random UUIDs. It's not clear exactly what the bug was
since the two versions have diverged enough, and the official version
has been spread across multiple files. However, it was easy to test
that the old copy would generate colliding UUIDs very easily, only
requiring a few tens of thousands of UUIDs before a collision would
occur (it likely has to do with timing since that's one of the
components used to seed the random UUID).