Fix ProxyObject hash to be sane.
Use boost::hash_combine instead of just ^'ing together the component
hashes. This turns out to be very important for a common case. Just
using xor is only kind of safe if the two hash values are
uncorrelated. Unfortunately, the two parts of the has for a
ProxyObject are its owner and the object identifier. Since we always
generate self proxies, these are definitely positively correlated for
your average ProxyObject. This ended up triggering some bad behavior
in a hash_map in the Emerson plugin when we ended up with a lot of
presences, each with their own self proxy, all put into one
hash_map. Since the two hash values are equal, the new hash value was
always 0 and everything ended up in one bucket.

Moral of the story: stop trying to do proper hash combining yourself
with ^.