Fix check for validity of redirect message.
Before processing a redirect message in the SessionManager, we check
to make sure it's actually valid and applies to our current request
since we might end up getting something out of date. However, the
check was assuming that if we had a connection request and the server
being connected to matched, that checking the seqno would be
ok. However, in the case of duplicate redirects, which can happen due
to high latency in the request triggering a retry, we might have reset
the seqno for the redirect step, but still be getting a connection for
that server, making the connectingTo server still the same. In this
case, we need to also check that the seqno is even still valid. Other
places we have similar checks aren't subject to the same issue because
redirects are the only time the connectingTo server and seqno are not
reset simultanously. This is related to the issue fixed in
dc0b5bff4786, i.e. they are both due to redirects only reseting the
seqno immediately but not the connecting server.