Change AggregateManager priorities to only use tree level.
We want the ordering to take into account the number of observers, but
the existing approach doesn't work since it can result in prioritizing
a parent above its child, making it impossible to make progress in
generating aggregates. This patch makes the prioritization depend only
on the level in the tree, but a better solution would address the
issue by prioritizing by observers under the constraint that all
children have to have a higher priority than their parents.