Improve query aggregation for solid angle queries.
Previously we just used the entire server region because it was
conservative. Instead, compute bounds around queriers. This needs to
be done periodically in case queriers are moving. Also, because the
merging algorithm isn't perfect, it's possible we could end up with
worse bounds. In that case, fall back to the bounding region of the
server as we did previously.