AggregateListener fix for libprox and better info from AggregateManager about failures.
Update libprox to get a fix for calls to AggregateListener that became
incorrect when changes to how notifications had to happen for
replication without lifting cuts. Also add better error reporting from
the aggregation method in AggregateManager so we can tell after the
fact why something is failing, especially useful when the item finally
fails for the last time.