Disable assertion that is being hit and can be safely avoided.
Its not clear why we're sometimes hitting this, but we can just bail
on the connection callback since it's just indicating a failure
anyway. It's likely due to some issue in SessionManager but it wasn't
clear how it could occur since the apparently correct checks are
performed before triggering the callback.