Use separate cpp files for each CXXTest suite.
Use the --part and --root commands to generate a separate file for
each test suite. This avoids recompiling all that code for a single
changed test. However, it also required a lot of fixes to missing
includes or misuses of namespaces because a lot of tests were relying
on other tests that were included before them either including files
or using namespaces.