Add SST test.
Rename TCPSST test classes to make room for SST classes. Add a Mock,
local datagram layer that supports adjusting delay and lossiness. Use
this to write a few simple tests of SST including basic connectivity
and send/receive over lossless/lossy and no delay/high latency
channels.

Currently some of these fail. This seems mostly to do with the
reasonable timeouts imposed to make sure the tests finish in a
reasonable amount of time, although with lossiness sometimes things
seem to stall. Its not clear if the stalling is due to a failure in
the code or just timeouts that are increasing.