Fix Connection service scheduling.
Fix the Connection service scheduling and document the approach and
reasoning throughout since it's kind of confusing and error prone. Now
servicing should only occur when there is something to do, either a
packet can be sent (one is available and there is congestion window
space) or a timeout needs to be processed. Various other fixes fall
out from this as well: we no longer incorrectly process initial
packets multiple times and therefore don't increase the number of
retransmissions incorrectly, and the number of retransmissions is only
used for timeouts during the connection phase.