When using strands, timers are always able to cancel (active) timers.
Fix the return value from IOTimer::cancel when using strands to always
return 1 since the ability to cancel the underlying ASIO timer doesn't
matter since we also track a value to skip the callback which is
guaranteed to work when using strands (as long ast the timer is
actually still active). Add a note to the method's documentation to
clarify the return value's semantics and when it's useful.

Also remove checking of this value in SST. Although it now returns the
right value, there's no point in checking it anymore since it will
always return 1.