Stop SST connection check alive callbacks when the connection is closing.
Convert SST connection check alive callbacks to timers and cancel the
timer when the connection is going to be cleaned up. This can happen
in a couple of ways, notably through cleanup() calls during normal
operation and close() calls that are invoked explicitly upon shutdown,
so we need to (possibly) cancel the timer in a couple of places.