Pass stop() requests to SST connections and streams.
Instead of calling closeConnections() in stop(), pass the call along
to Connections, which then pass it on to streams, giving them a chance
to cleanly shutdown. Streams are closed with force=false so they
should finish sending and cleanup cleanly. This, combined with
stopping the checkAlive timer, should also allow Connections to then
be cleaned up.