Warn about posting events with long timeouts.
Posting events with long timeouts is a bad idea since they cannot be
canceled. If a shutdown is requested, then they hold everything up
because they timeout long after the shutdown is requested. This just
logs a warning, with tag information, when such a post is performed to
encourage people to remove them.