Convert TCPSST timeout to use a cancelable timer.
Rework ASIOStreamBuilder into a class which tracks a queue of timeouts
and uses a single timer to cancel them. This gets rid of an
uncancelable post() with a long timeout that can hold up shutdown (and
which now generates a complaint in the log). It also has the nice
benefit of removing some completely unnecessary global state.