Make system tests track dict of binaries.
Instead of having separate parameters, which gets annoying with more
than just space and cppoh, use a dict which is filled in when we start
the tests. This is simpler, makes it easier to expand to all the
binaries (this addds pinto and cseg), and avoids duplicating the code
to get the full path to the binaries in each of the test classes.