Reorganize system tests so there is a generic "WorldHttpTest".
Move the code for ProximityTest into a generic WorldTest in the
framework. This can serve as the basis for a world which is driven by
HTTP requests to the services. We were already reusing ProximityTest
in the aggregates tests for this purpose. This makes things more
generic and allows us extend and reuse it for other, more complicated
tests beyond what the existing proximity tests need, e.g. to support
multiple servers as well as the single server case.