Only ack packets once we know their contents are fully processed.
In some cases, specifically packets with data, we may fail to actually
accept the full contents of the packet. If we start using the packet
sequence number for acks immediately, we may end up acking a packet
which we haven't actually saved the data from.

To fix this, only update mLastReceivedSequenceNumber at the very end
of receiveMessage() if it was actually fully handled. Currently we can
safely assume all packets *except* DATA packets will be handled
properly. Other packets (INIT and REPLY) can contain data, but we
don't need to worry about them because the maximum amount of data on
the first packet (the only case for INIT and REPLY) is smaller than
the minimum buffer size. We return a bool from the handling method to
determine whether we can ack that packet.

However, because some acks and packets are generated before we finish
processing, we may want to make sure we use the newest value as soon
as we possibly can. Therefore, change sendData() to take the sequence
number that should be acked so we can specify it explicitly if we have
a newer one than the last one recorded so far. The original version,
using mLastReceivedSequenceNumber can also be used, but is now named
sendDataWithAutoAck, indicating we'll automatically add the last
received packet's sequence number. This should really only be used
when sending a packet *not* in response to an incoming packet.