Use simple tracking of segments in SST rather than a bitmap.
Switch from a bitmap tracking valid segments of received data on an
SST stream to a simple list of ranges. Updating this region is a bit
more complicated (since it isn't just a memset), but getting the next
valid range of data is simple and this doesn't eat up so much memory (we were
doing even worse than we had to with 1 byte / byte rather than 1 bit / byte).