Fix bug in tracking valid SST buffer ranges.
When inserting an element into the front of the list, the length
wasn't being updated to include the new data. Upon drops/reordering,
this could cause stalls if the retried/reordered segment was at the
beginning of the buffer since this caused a segment of that size to
effectively be lost from the end of the buffer.