Fix invalid use of temporary references.
48d1c334a646fbc5f446221150d349796c1e6578 pulled a temporary out of
some notification calls to support LLVM compilers, but only copied out
some lines where a reference to a temporary was being used. Previously
it worked because the temporary lasted the lifetime of the call, but
when separated it was no longer valid through the notification
call. Just separate it out as a named variable instead so it has the
proper lifetime.