Add per-node data containing bloom filters for Twitter terms.
Add per-node data classes for query processing that are the same as
bounding sphere + max object size, but also decode and aggregate bloom
filters. The main implementation is in libtwitter, but separate
classes inherit from it to support the space and OH because some
options need to be specified, but come from different places for the
two binaries. Use these new classes to register a few new handlers
that allow us to perform manual querying on the space server and rtree
queries on the client. The client option is currently only used as a
baseline since there isn't yet a query handler to take advantage of
the extra data.