Fix MigrationMonitor to handle timeouts when objects are still in the region.
The MigrationMonitor is designed such that objects must always have a
next time to an event. If they aren't moving, this time should be
infinite. Since we set a fixed timeout, this fixes the timeout handler
to check one more condition -- that the object isn't still on the
correct server -- before invoking the callback which triggers the
migration.