Check that ClutterRenderer actors are valid before using them.
When ClutterRenderer handles an Invokable callback, it was blindly
using the result from get_actor_by_id. However, that can return NULL
if the actor with the specified actor ID isn't found. This change
follows all calls to get_actor_by_id by a check which ignores the call
and returns false if the actor isn't found.