Don't log invalid loc update messages.
Location updates aren't framed, so we don't know when they are
complete. Therefore, we already have support for retrying the parse
after getting more data if the first attempt to parse failed. However,
we were logging invalid messages if we couldn't parse them, but that
is much more common now as updates with extra query data can easily be
larger than the 1k default SST packet size. This creates too much
console spam. The lines are left in there in case they are handy for
debugging.