Add target result size control to TermRegionQueryHandler.
Add a target number of results parameter to
TermRegionQueryHandler. There are better ways to implement the
control, but currently it uses a simple feedback loop, adjusting a
minimum bounds size parameter when the results grow too large. To
allow the results to expand, it waits awhile and then starts reducing
that size so that more objects will be admitted. Effectively, this is
just dynamic control by the system of a query parameter with the goal
of matching the target number of results. It works well enough for now
to be usable.