Provide leaf object notifications for AggregateManagers.
Add methods to AggregateManager which are invoked when leaf objects
are added/removed from the query processor. These come from the same
thread/strand as the other aggregate events, making it possible to
trivially reconstruct a tree. This requires an updated version of
libprox which provides these events via
Prox::AggregateListener. Update TwitterAggregateManager to take
advantage of this. It now is more flexible than it has to be since it
previously was mostly OK with events from multiple threads, but
shouldn't have threading/consistency issues anymore.