Make TwitterAggregateManager handle changing parent/child relationships.
Sometimes, while an aggregate is being computed, its parents can
change. We collect the list of parents before processing to make sure
we have a complete list, but then we still assumed all parents from
the start of processing were still dirty and needed to be
enqueued. However, if the last dirty child is removed while it's being
processed, then this won't be true: the now dirty-child-free parent
can immediately be processed, and could even finish before the
child. This fixes that handling to verify the parent-child
relationship still exists before processing the parent.