Fix bug in cleaning up removed chlidren in TwitterAggregateManager.
Fix issue where sometimes a parent aggregate's dirty_children entries
were not being cleaned out properly when the child was removed. This
would occur if the aggregate was being processed just at the time the
removal happened because we only did the operation if the dirty flag
was set on the child. However, since it could be marked clean but
still be processing, we need to check that flag as well.