Pass root aggregate query data updates to PintoServerQuerier.
When the root aggregate's query data is updated, make sure it is
passed to PintoServerQuerier. This requires a new method in the
interface and support from the proximity implementations. This also
updates the PintoServerQuerier implementations to pass this
information along to the Pinto server. The local implementation just
stores it and includes it in updates, allowing the top level Pinto to
work naturally. The real Pinto doesn't yet handle extra query data,
but the PintoServerQuerier will pass it along so it can in the future.