Fix incorrect parsing of old-style solid angle queries.
Fix a bug introduced by 371d983ed in parsing old-style solid angle
queries. To support flexible, custom parsed queries, we needed to
indicate whether parsing was successful. The test was supposed to
check that at least one field of the classic queries was included, and
default to classic queries. Instead, it was checking for all of them,
making it fail, e.g., if max_results was not provided. This fixes that
condition and also handles default values properly for these queries
when the parameters are not provided.