Make coalescing first set of results optional, disable by default.
Add an option to the prox plugin to make coalescing the first set of
declarative (non-manual) query results optional. Coalescing them
sounds nice because it gets rid of any intermediate steps that need to
be reported and dumps all the data to the client at once. However, it
means we need to wait for the initial query evaluation to finish,
which could take a bit when the tree is large. Disable coalescing by
default.