Handle (un)subscription requests in AlwaysLocationUpdatePolicy more efficiently.
Instead of passing all requests with post(), using the main strand as
an implicit queue, explicitly store requests, handle them quickly if
they reverse a still-waiting earlier request, only post requests for
processing when one isn't already running, and handle them in batches
so there isn't so much overhead to do minor modifications to the
subscription data structures. The tradeoff is that we now do more
checks up front in the query processing thread. However, this should
improve efficiency in one particular case where we block up the main
strand by posting so may tasks that take a long time to process when a
query is first evaluated and all parts of the tree up to the cut need
to be added and immediately removed.