Longer initial RTO and logging.
Make the initial RTO for connections in SST longer by just multiplying
the value. This makes sure the initial value isn't too low an
approximation if the first round trip happens to be very quick due to
low load (both network and CPU). If it remains as low as the initial
value, it will quickly converge anyway.

Also add a couple of logging statements at insane level for experiments.